feat(erratic-lemon): refactor forward of ui
- align with new options
- use templateUrl
- move other components to their files
- install lightningcss
- update title and icon
