+++ packages/erratic-lemon/package.json
@@ -12,7 +12,14 @@
     "@microphi/flux": "^2.7.3"
   },
   "devDependencies": {
+    "lightningcss": "^1.28.1",
     "typescript": "~5.6.2",
     "vite": "^5.4.10"
+  },
+  "nvim": {
+    "pde": {
+      "styles": [
+      ]
+    }
   }
 }
