+++ packages/erratic-lemon/src/ui/components/components/fx-if.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@microphi/flux';
+import { Input } from '@microphi/flux';
+
+@Component({
+  selector: 'fx-if',
+  template: `
+    <slot style="display: {{condition}}"></slot>
+  `
+})
+export class FxIf {
+  @Input() condition!: boolean;
+}
