+++ packages/erratic-lemon/src/ui/components/components/fx-item.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@microphi/flux';
+
+@Component({
+  selector: 'fx-item',
+  template: `
+    <slot></slot>
+  `,
+})
+export class FxItem {}
