+++ packages/erratic-lemon/src/ui/components/fx-root/fx-root.component.html
@@ -63,5 +63,5 @@
 <h1>Test Cases</h1>
 
 <test-component-sh firstname="{{name}}" lastname="{{name}}">
-  <h4>transcluded name {{name}}</h4>
+  <span>transcluded name {{name}}</span>
 </test-component-sh>
