--- packages/erratic-lemon/src/ui/external.d.ts
@@ -1,8 +1,0 @@
-// external.d.ts
-// eslint-disable-next-line @typescript-eslint/prefer-namespace-keyword
-declare module JSX {
-  type Element = string;
-  interface IntrinsicElements {
-    [elemName: string]: any;
-  }
-}
