--- packages/erratic-lemon/src/ui/fx-for.component.ts
@@ -1,40 +1,0 @@
-import { Component, Input } from '@microphi/flux';
-
-
-@Component({
-  selector: 'fx-content',
-  template: 'this is fx-content'
-})
-export class FxContent {
-
-
-  constructor() {
-    console.log('constructing fx-content');
-  }
-}
-
-@Component({
-  selector: 'fx-for',
-  template: `
-    <h1>{{cursor}}</h1>
-    <h1>{{list|json}}</h1>
-    <fx-content></fx-content>
-  `,
-  shadowRoot: true
-})
-export class FxFor {
-
-  d = console.log;
-
-  @Input() cursor!: string;
-  @Input() list!: any[];
-
-  constructor() {
-    this.d('fxfor component', this);
-  }
-
-  fxOnChanges(changes: any) {
-    console.log({ changes });
-    this.d({ changes });
-  }
-}
